Basic turn functions implemented
Added handTotal function to calculate the users current card total.
Added isBustCheck to see if the player or computer has gone bust.

Added the mainGameLoop, playersTurn and dealersTurn functions to layout
the turn structure of the game. Currently you can complete one turn but
their are several bugs and no blackjack check.

Laid out a checkForBlackjack functions, but is not implemented yet.